Move `Vec` into a submodule
