+++ src/vec/raw.rs
